aggiunte migliorie per quanto riguarda lo stile delle pagine e alcune funzionalita

+++ src/main/java/it/prova/myebay/service/AnnuncioServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import it.prova.myebay.exception.AnnuncioChiusoEccezione;
+import it.prova.myebay.exception.UtenteNotFoundException;
 import it.prova.myebay.model.Annuncio;
 import it.prova.myebay.model.Utente;
 import it.prova.myebay.repository.annuncio.AnnuncioRepository;
@@ -43,6 +44,18 @@ public class AnnuncioServiceImpl implements AnnuncioService {
 	@Override
 	@Transactional
 	public void aggiorna(Annuncio annuncioInstance) {
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Utente utenteFromDb = utenteService.findByUsername(username);
+
+		if (utenteFromDb == null)
+			throw new UtenteNotFoundException();
+
+		if (!annuncioInstance.isAperto()) {
+			throw new AnnuncioChiusoEccezione();
+		}
+		annuncioInstance.setAperto(true);
+		annuncioInstance.setDataCreazione(LocalDate.now());
+		annuncioInstance.setUtente(utenteFromDb);
 		annuncioRepository.save(annuncioInstance);
 	}
 
@@ -53,7 +66,7 @@ public class AnnuncioServiceImpl implements AnnuncioService {
 		Utente utenteFromDb = utenteService.findByUsername(username);
 
 		if (utenteFromDb == null)
-			throw new RuntimeException("Elemento non trovato.");
+			throw new UtenteNotFoundException();
 
 		annuncioInstance.setUtente(utenteFromDb);
 		annuncioInstance.setAperto(true);
@@ -64,9 +77,16 @@ public class AnnuncioServiceImpl implements AnnuncioService {
 	@Override
 	@Transactional
 	public void rimuovi(Long idAnnuncio) {
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Utente utenteFromDb = utenteService.findByUsername(username);
 		Annuncio annuncioFromDB = this.caricaSingoloElemento(idAnnuncio);
+		if (utenteFromDb == null)
+			throw new UtenteNotFoundException();
 		if (!annuncioFromDB.isAperto()) {
 			throw new AnnuncioChiusoEccezione();
+		}
+		if (annuncioFromDB.getUtente() != utenteFromDb) {
+			throw new RuntimeException();
 		}
 		annuncioRepository.deleteById(idAnnuncio);
 	}
@@ -75,6 +95,18 @@ public class AnnuncioServiceImpl implements AnnuncioService {
 	@Transactional(readOnly = true)
 	public List<Annuncio> findByExampleRicerca(Annuncio example) {
 		return annuncioRepository.findByExampleRicerca(example);
+	}
+
+	@Override
+	public Annuncio caricaElementoConUtente(Long id) {
+		return annuncioRepository.findByIdConUtente(id).orElse(null);
+	}
+
+	@Override
+	@Transactional(readOnly = true)
+	public List<Annuncio> gestioneAnnunci(String username) {
+
+		return annuncioRepository.findAllByUtente_Username(username);
 	}
 
 }
